requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
Django==3.2
django-cors-headers==3.8.0
django-environ==0.7.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
future==0.18.2
html5lib==1.0b10
httplib2==0.20.4
idna==3.2
Pillow==9.0.1
psycopg2-binary==2.9.1
pycparser==2.20
pyparsing==3.0.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.1
reportlab==3.6.0
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
stripe==2.60.0
urllib3==1.26.7
webencodings==0.5.1
xhtml2pdf==0.2b1
